+++ scripts/tests/test_all.sh
@@ -36,6 +36,8 @@ echo "--- Test 16 ---"
 ./scripts/tests/test_bongocat_16.sh
 echo "--- Test 17 ---"
 ./scripts/tests/test_bongocat_17.sh
+echo "--- Test 18 ---"
+./scripts/tests/test_bongocat_18.sh
 echo "--- Test 20 ---"
 ./scripts/tests/test_bongocat_20.sh